rename base.nix to default.nix
I previously split these into two files because I needed to have an expression
that evaluated to a function and a different expression that could be the target
of `nix-env -rif`. But it turns out nix-env is perfectly happy to take an
expression that evaluates to a function with a default argument, and then call
that function. So these do not actually need to be separate files, I can
accomplish both goals with the same file.