undo sourcing some bash completions
The completions for exa, and possibly others, require the
bash-completion package, which appears not to be compatible with
bash 3.2, which I have on my mac. Revert for now.

I should probably find a way to upgrade my shell, and possibly
experiment with fish. It would be nice to manage my shell through nix as
well.