copy in exa derivation
The last release of exa was about 18 months ago. Since then there have
been a lot of changes, including one that I care about in particular:

    9bb62fbd631090ae9b24782c594506266a1e47bb:
    "make exa on symlinks to directories show their contents"

I've been using nix heavily recently, making many modifications to my
environment and examining the effects. Nix uses symlinks all over the
place, so this means I often want to list the contents of a symlink to a
directory. Sadly, the latest release of exa does not do that. But the
latest version of the code does! So I am going to manually edit the exa
derivation to build from a more recent version of the code.

There is probably a more succinct way of doing this, but I do not know
what that way is. So what I'm doing is just copying the whole exa
derivation from nixpkgs, and then I'll change it to pull a later version
of the code.