mechanism for one-off scripts
I wanted to create a shortcut to execute `rg '\bSOME-PATTERN\b'`. A tiny
shell script is the best mechanism I know for this. I did not yet have a
mechanism for adding shell scripts to my nix environment, so I created a
general-purpose mechanism for that. I also usually avoid scripting in
bash, and knew that for future scripts I would want to be able to use
other languages, too. So I also created the same mechanism for several
languages I might want to use in the future.

This is a lot of overkill for the one script I wanted to add, and is at
some risk of generalizing too early. But writing little scripts like
this is a thing I did a lot before I started using nix, and the
probability of me wanting to continue doing this in nix seems relatively
high.