remove an obsolete bit of bash config
Before nix, I used to manage my dotfiles and some of my dev environment
with vcsh. With vcsh, I had to remember to commit and push any changes I
made to my local machine, otherwise the dotfiles repo would get out of
date. (With nix, this step is already built into the workflow by way of
the local config files being immutable). To remind myself to do that, I
added a message in my prompt whenever my local environment was out of
sync with the dotfiles repo. (A message which I often ignored, which was
a separate problem).

My usage of vcsh is now thoroughly in the past. Nix has been working
very well for me, and I don't plan to go back. This vcsh message has
been disabled for a while, and it will probably never be re-enabled, so
now it's gone.