move vim-config.vim to xdg config directory
This makes it a little more like my other custom configs. Sadly, it's
still a special snowflake in that it is not picked up by setting the
XDG_CONFIG_HOME variable. Instead, I'm still using the `customRC` field
in the neovim package, which seems to be the only way to customize
init.vim using that package.

This makes the vim config patchable in the way that my other configs
are.