configure my shell in kitty rather than sys config
I install a custom shell in this nix environment, which has nice
benefits over using the system-default shell (and over using the
system-default shell with a custom config). See these commits for more:

    f2161d6c6922a2f9033b01411ffda4ab993d48b2

    and

    cb85564031bc5f00eedccfa0276c415246012dd1

While it was quite nice to take this incremental step toward having more
of my environment checked as a nix expression, this still left me with a
bit of external system-dependent config to manage: telling the OS to
make `shell` my login shell.

That extra bit of config was a bit ugly. But it also made my (private)
NixOS config much harder to reproduce, because it introduced
co-recursion between the system config and the user config.

This change simplifies the whole thing nicely.