kitty: XDG_CONFIG_HOME -> KITTY_CONFIG_DIRECTORY
My typical way to set configuration options for programs installed
through nix is by automatically setting the XDG_CONFIG_HOME env var when
I launch the program. This works well except when the program is my
terminal (or, in principle, any program which launches other programs).
In that case, every program I run inside my terminal will _also_ have
that variable set, which is not what I want.

This commit fixes the problem for kitty, by setting a kitty-specific env
var. Note that this variable needs to point directly to the kitty config
directory, wherease the xdg variable I was previously setting needed to
point to the directory _above_ the kitty config directory.