vim: fix config after 0f773aa06befb6ed18d5644afbe
Turns out that vim was depending on reading the XDG_CONFIG_DIR that was
being set by kitty. This is not the behavior I actually wanted, and it's
now broken.

This commit embeds the config file directly in the package that nix
installs, which is the behavior I wanted all along.