bump nixpkgs to 20.03
As of now, magic-wormhole in 20.03 has a transitive dependency on
OpenSSL 1.0.2. That version of OpenSSL is marked "insecure" in 20.03,
because it is no longer maintained upstream. nix-env refuses to build
insecure packages by default.

I tried to manually whitelist OpenSSL 1.0.2, but this resulted in
nix-env trying to build it from source. No one has time for that, so I'm
commenting magic-wormhole out of my config for now. Hopefully I can add
it back soon-ish. It's a nice tool (that I only use occasionally).