bash: fix completions for my 'g' alias of git
After upgrading to nixpkgs 20.03, but before making this change, I was
seeing this error message when starting a new shell:

    shell: __git_complete: command not found

And when I tried typing: `g a<tab>`, I wasn't getting completions.

I'm not really sure what caused this to break. I found this issue when
Googling the problem:

    https://github.com/scop/bash-completion/issues/278

It seems related. In it, the maintainer of bash-completions says that
his package never did ship __git_complete.

When looking through my history of changes to completion config, I saw
f5338a6ce64003f29bfd7477dc6d4e5241754ab8, which mentions that
$NIX_PROFILE/share is full of bash completions that I am not sourcing
because it is too slow. (But which nonetheless are apparently loaded by
bash-completions anyway, asynchronously.) I checked if there was a git
completion in there that defined __git_complete, and sure enough there is.
Sourcing it fixes the problem.