bash: put nix tools in PATH before using them
Turns out I was depending on this bashrc being sourced after
~/.nix-profile/bin was put into PATH, which caused error messages when I
tried to use `shell` in iTerm2.

In principle I could find a way to reference the relevant tools
(restore_colors and direnv, here) by their absolute path in the nix
store. But it's not clearly worth doing so yet, so I'm not.