re-enable magic-wormhole, and bump python to do so
The current version of magic-wormhole on nixpkgs 20.03 (which is the
latest stable channel, and the one I'm using for most of my packages)
has a transitive dependency on openssl 1.0.2u. That version of openssl
is no longer supported, and is therefore marked as insecure in the
corresponding nix derivation. Packages marked insecure do not build by
default, and since that openssl version would not build, magic-wormhole
would not build either.

It was possible to enable the build by explicitly opting in to ignoring
the insecure flag on that openssl version, and I manually did that for a
little while.

But the better fix is to use a version of the magic-wormhole derivation
that uses a more recent, supported openssl. The latest version on the
unstable channel does that, so I'm switching my whole python + python
package installs to the unstable channel. When 20.09 is released, I
expect I will switch them back to stable.