alias "vit" to open vim w/ terminal panel on right
neovim has a built-in terminal emulator, so you can use a terminal
within vim. I like having that terminal open in the right have of my vim
window, with code in the left half, like so:

---------------------------------------------------
|                        |                        |
|  code code code code   | $ <terminal stuff>     |
|  code code code code   |                        |
|  code code code code   |                        |
|  code code code code   |                        |
|  code code code code   |                        |
|  code code code code   |                        |
|                        |                        |
---------------------------------------------------

Usually when I open vim I just want the text editor. But there are many
times when I want the terminal, too.

I can enter a few commands within vim to open the terminal, but it's
quicker to instead launch vim with the relevant flags to cause it to
open the terminal, in my preferred location, with my preferred shell, as
a single command.

I figured out the relevant command a while ago. Since then, I have been
using ctrl-r to find it in my shell history whenever I want to use it.
But today, it didn't come up immediately in my search, because I have
been running some similar-looking but unrelated commands for the past
couple days. At that point, it seemed to me that this command deserved
its own shorthand, so I decided to give it the alias "vit", for
"vi (with) terminal".