better derivations for hasktags and hoogle
This commit wraps both hasktags and hoogle with
"haskell.lib.justStaticExecutables". That wrapper does what it sounds
like: it yields only the static executables from the corresponding
packages. It drops the dylibs and whatever else they were installing.

One reason this is better is that it decreases the size of the closure
that is installed in my profile.

This also decreases conflicts: most of the haskell packages I have tried
to install have included a .dylib for ghc, at the same path in the
profile, which requires me to manually and arbitrarily choose one to
prefer. Using this wrapper drops that .dylib, removing the conflict.

There is a downside to using these wrappers: the new derivations are not
in the binary cache, so I have to build the haskell executables. This
doesn't take super long, because all of their dependencies are still
cached, but it still adds to the build time. This cost seems worth it
for now, but I may change my mind about that in the future.

There is also the downside that the wrapper is a bit ugly and the code
in default.nix is now longer. Alas.