use fzf for ctrl-r
This used to work a long time ago. It may have broken when I switched to
using nix. I switched laptops and operating systems at the same time, so
it's very plausible that I would have not noticed this after spending
many days or a few weeks slowly rebuilding my dev environment.

This config broke because it used a file, ~/.fzf.bash, that I did not
have checked into my dotfiles repo. Happily, it was easy to fix. The fzf
derivation in nix includes a file that does the same thing. I just had
to get the right path for it.

fzf is a nice ctrl-r interface, and I've missed it. It's good to have it
back.