bash: better history search on ctrl-r
In the previous commit, I switched to using fzf for ctrl-r. This
improved the search UX, but it still searched over the same corpus:
~/.bash_history.

This commit changes ctrl-r to search through my ~/.full_history file
instead. That file, which I've been maintaining for 9 months now, stores
all of my commands. It does not get truncated like ~/.bash_history does.

I was motivated to make the switch after I used ctrl-r to search for a
command I knew I had run not that long ago, only to find that it had
disappeared from my ~/.bash_history.

This change was more involved than I expected. It turns out that the
format I was using for history entries in ~/.full_history was not as
machine-parseable as one would hope. All of the fields were separated by
spaces, and some of the fields had internal spaces, and there was no way
to distinguish the two uses of space aside from using my priors about
the kinds of commands I would enter. To solve this, I changed my history
logging to use logfmt, which is a little more verbose but entirely
unambiguous. This also required me to put a custom reader in front of
fzf's default history search, to parse out of logfmt.

This commit also includes convert-hist.go, which I used to convert my
existing record to the new logfmt version. I only needed it once, so I
will delete it in the following commit, leaving it here just for the
record.