new pattern for config files
This repo contains config files for some of my tools. These files end up
getting copied into the nix store. In order to actually use them, the
tools need to be pointed to these files somehow.

My previous solution was to point them via an environment variable. One
obvious way to do this was to set $XDG_CONFIG_HOME to
$NIX_PROFILE/config, system-wide. Setting it system-wide turns out to be
a problem because, among other reasons, some tools expect to be able to
write to their config directory, but anything inside the nix store is
read-only.

To work around the problems with setting $XDG_CONFIG_HOME system-wide, I
instead set it per-tool. nix provides a utility to wrap programs in a
little shell script wrapper that lets you set custom environment
variables. This way I could opt in to having specific tools see the
$XDG_CONFIG_HOME var pointing to the nix store.

(Aside: Some tools don't read $XDG_CONFIG_HOME, but do read some other
variable eg $KITTY_CONFIG_DIRECTORY or $STARSHIP_CONFIG. This has turned
out to be a minor point that is easy to accomodate so far.)

That was how my config worked until yesterday.

But it turns out there are some problems with that approach. I ran into
two of them, which could both be described: "what if the tool gets used
without going through the nix wrapper?"

Case 1: starship

I added starship recently: 9610f9e13a7e7d8d7222f9ba25633e2d2e8260f5

I had to set its config var system-wide, rather than use the usual
wrapper. Why? Because of a leak in the abstraction of the wrapper
script. It went like this: I added the following line to my bashrc:

    eval "$(starship init bash)"

`starship` there resolves to the wrapper script, which would set the
environment variable the way I wanted. But the command
`starship init bash` just prints more shell script, which then has to be
eval'd. And the shell script it prints _includes a reference to
starship_, and that reference _does not resolve to the wrapper_, it
resolves directly to the underlying binary. That's because
`starship init bash` uses $0 as the path to starship, and $0 is the
underlying binary, not the wrapper.

Case 2: the git_status module in starship

starship is a nice prompt. By default it prints info about your git repo
status, whenever the current working dir is inside a git repo. It uses
libgit2 to do this, rather than invoking a git binary. This is a great
idea -- it is more efficient and more robust. But of course the library
knows nothing about my wrapper script. There's no way it could! And
because of this, the git status line in my prompt was using a different
config from the git command I ran in my terminal, and showing different
results accordingly.

I considered trying to make a wrapper for starship that set
$XDG_CONFIG_HOME, _just for the git_status module_, but decided against
it. It seemed surprising for the prompt to be using a different env
config from stuff I could type after the prompt. (And I don't know how I
would even accomplish that setting anyway, given the problem in case 1
above).

So the solution I am trying now is: everything just lives in ~/.config.
This is what all of my tools expect by default, and it's pretty simple.

The problem with this, and the reason I wasn't doing it already, is that
(as far as I know) nix-env can't touch ~/.config, by design. All it does
is add a bunch of symlinks under ~/.nix-profile.

So to modify stuff in ~/.config, I decided to add a post-install hook
for my nix config. I'm a little unhappy about having to do that, but it
solves this problem nicely and also gives me a way to solve another
problem that has been outstanding for a while now: nix-env _already_
didn't install everything I needed. I currently have some English text
in the README.md in this repo explaining how to do the rest.
Installation through English text in a README is the worst way to
install something. Now that I have a post-install hook, I can convert
that English to code.

This commit adds the post-install hook as bin/go/jeremy-post-install.go.
Like anything else in the bin directory, this will automatically get
installed in the $PATH as `jeremy-post-install`.

This commit also modifies the `i` alias, which I use for re-installing
my environment, to run `jeremy-post-install` after `nix-env`.

This commit also remove the shell wrappers and the system-wide
$STARSHIP_CONFIG env var.

We'll see how this goes.