temporarily disable iterm2 color controls
This is broken on the latest version of iTerm2.

The latest version of iTerm2 requires an updated version of the iterm2
python library. That updated library requires pyobjc. The pyobjc
derivation on nix is currently broken. Until someone (probably me) fixes
the pyobjc derivation on nix, I won't be able to use my nice little
color-changing script :'(