fix bug in flake port (symlinkJoin -> buildEnv)
Before I switched to flakes, I was using a default.nix file. An
expression in a default.nix file can yield a list of derivations rather
than a single derivation, and nix-build would handle merging them for
you.

The default package in a nix flake, on the other hand, must be a single
derivation. So I have to do the merging myself. When I made the port to
flake.nix, I knew of the "symlinkJoin" builder, and it seemed to do the
trick, so I went with it. But it turns out not to be sufficient to
replace nix-build's builtin merge logic. I didn't notice for a while,
but today I saw one of the symptoms: despite the fact that I merge the
"go" derivation into my environment, I was not getting the go binaries.
I don't understand why not, though the docs for symlinkJoin do say:

    BEWARE: it may not "work right" when the passed paths contain symlinks to directories.

I did a brief search for an alternative and found buildEnv. It seems to
be designed for the purpose of doing these merges for a profile, and it
solves the dropped go binaries, so I'm switching to it.

NOTE: One thing buildEnv does not do, which nix-build and symlinkJoin
both do, is automatically flatten its input list. So I do that
flattening myself in my "bundle" function, since I rely on flattening
for "my-scripts".