starship: make it more obvious when I'm using ssh
The default starship config has two modules that render only when I am
using ssh: hostname and username. When I'm on a local machine, my prompt
currently looks something like this:

    ~/nix at 22:14:08 ❯

And when I'm in an ssh session, the prompt looks different:

    jeremy on linux in ~/nix at 05:15:09 ❯

It's certainly longer, but "jeremy" and "linux" are rendered in somewhat
muted colors, and I found myself in practice not noticing the
difference. The size of the prompt aso varies in normal use as I switch
directories.

I find it helpful to have a clear visual distinction between shells on
remote machines and shells on my local machine. Otherwise it's easy to
forget what machine I'm talking to! Especially when I have a bunch of
terminals open, some local and some remote.

So this commit changes the hostname (which, again, renders only when I'm
on ssh) to a more vibrant color, sets it in italics and underlines it,
and adds a cloud emoji that stands out against the rest of the plain
text.