starship: less obnoxious ssh prompt
In "6b837ee6 starship: make it more obvious when I'm using ssh", I made
my prompt have a visually loud indicator when I was connected to a
remote machine over ssh.

Now that I've used it a bit, it's too much. This commit makes the
indicator a bit quieter and a bit less ugly: it removes the underline
and adds a space between the cloud icon and the hostname.