git: suppress new warning in git 2.27
git 2.27 [added a new warning][1] when pulling without an explicit
configuration of whether to rebase or merge on changes.

The justification provided is that "often noice Git users... end up with
an unnecessary merge from upstream" when pulling in changes.

I'm a bit confused about this justification. In my experience, git has
always refused to update my local branch when I pull and there are
conflicting changes, and then it gives me some messaging about what to
do. I think the default messaging led me to re-doing the pull in such a
way that I ended up with a merge, as the justification says happens,
until I learned that I could do "git pull --rebase" instead. But if that
is true, and we think novices should be doing "pull --rebase" more (I'm
a fan), why not change the messaging there, when it happens, to suggest
"pull --rebase"? Instead, git now warns you any time you do any pull at
all without an explicit configuration for this situation that only comes
up in some pulls.

I'm also confused about the suggested configurations. They suggest one
of:

    git config pull.rebase false
    git config pull.rebase true
    git config pull.ff only

"pull.rebase true" sounds like it is the same as always using the
--rebase flag. That makes sense, and is almost what I want, except that
I like getting interrupted when a rebase is necessary. I almost always
want to do the rebase anyway, but this informs me to update my mental
model at the same time, which I would not do if I didn't know it was
happening.

My confusion is about "pull.rebase false". It seems like this
automatically creates a merge commit? But the text also says that it was
the default? But that is behavior I have never seen, despite not
touching the configuration before?

I'm confused about something here.

But it seems most likely that "pull.ff only" is the setting I want, so
I'm going with it.

(Wanted because it will not create unneeded merges, and because it will
not rebase silently without nudging me to update my mental model first.)

[1]: https://github.com/git/git/commit/d18c950a69f3a24e1e3add3d9fc427641f53e12b