add-hist: fix data race in writing history lines
In rare cases, some of the lines in my history file will be interleaved.

I think this happens when multiple processes write to the history file
at the same time.

To prevent this, I have added a file lock around the write operation. I
rearranged the code a bit to minimize the amount of code under lock, and
I also changed the control flow for exiting on errors to ensure that
defer'd cleanups always happen.