fzf-hist: better behavior in error condition
fzf-hist is part of my implementation of command history search.

Its job is to read my custom history file, parse it, deduplicate the
commands within, and print them out in a format ammenable to search.

When the history file is not formatted as expected, it exits early with
a non-zero exit code. Prior to this commit, it also printed the
offending line. It turns out that this is not useful to me when I see it
in the course of attempting to do a history search on the command line.
Typically, I see a flash of "bad line:" and then a blank search box.

When my history file is malformatted, I want to fix it right away. And I
want this error message to help me do it. By simply printing out the
line number where the failure happened, I get a straightforward error
message when the problem presents itself, and I can find the issue
quickly in my editor.

In the previous commit (dd23e42 add-hist: fix data race in writing
history lines), I think I already solved the only problem that caused my
history file to become malformatted in the first place. So I might not
ever hit this error condition again. But maybe I will. If I do, I would
rather have this new error message than the original.