vim: generalize my sorting macro
This is one of the oldest parts of my config. I wrote this sorting macro
back when I was at Google, some time between 2011 and 2013. It's been
with me ever since, and I haven't updated it in all that time.

I used it a lot when I was writing C++, to keep my include statements
sorted, but I use it much less often now that I'm mostly writing in
other languages. Notably, Go has goimports, which sorts import
statements and does a lot more.

I tried to use the macro today to sort Python imports, and it did not do
what I wanted. I had a block of lines at the top of the file that I
wanted to sort, and instead it sorted a much longer section of the file.
The problem was that it assumed that the import statements had vertical
whitespace preceeding them, and failed somewhat catastrophically if that
was not true.

I've updated it to work if the import statements either have vertical
whitespace above them _or_ if they run all the way to the top of the
file. The new version works nicely.