i.py: depend on nixFlakes
This is a bit of a hack to solve this problem:

I have a NixOS machine with a system-wide install of nix that does not
support flakes.

I do not have a good way that I know of to replace that system-wide
install with a version that supports a recent version of flakes.

But I _can_ install my scripts with arbitrary nix package binaries on
the path. So my solution is to have i.py see a recent version of nix,
which I use solely for running this script, and which is different from
the version of nix on my system.

In principle it seems like this might cause conflicts, using two
different versions of nix, but I'm betting that in practice it will be
fine.