bash: delete obsolete python alias
In Python 2, it was possible to cause problems by mixing tabs and
spaces. The -tt flag issued an error if you used tabs and spaces in a
bad way. In other words, it turned on a strict linter to catch problems
you would otherwise miss.

In Python 3, mixing tabs and spaces at all is an error, and the -tt flag
is a no-op.

This alias causes me to set the -tt flag automatically whenever I type
"python" at the command line. But "python" now always refers to a Python
3 interpreter for me, and so setting this flag is always redundant.

So now I get to delete it and have a slightly cleaner config :)

---

References I found while confirming that my picture above was correct:

StackOverflow answer about -tt:
    https://stackoverflow.com/a/25054402

The exact line of code in the Python 3 interpreter that ignores any -t
shortflags:
    https://github.com/python/cpython/blob/c322948892438a387d752ec18d1eb512699a4d67/Python/initconfig.c#L2002-L2004