colorscheme: comment out broken iTerm2 logic :(
iTerm2 has updated and is no longer compatible with the old version of
the API library I was using. The new version of that library doesn't
build on Nix yet, because it depends on pyobjc, which does not build on
Nix yet because it's pretty complicated and no one has put in enough
work to get it done yet.