bash: don't complain if bash_completion.d missing
At the moment, I only have a single package that adds a completion to
this directory, and that package is in one of my configs that is not
even public.

When I install my environment without that config, this
bash_completion.d directory does not get created, and this bit of my
bashrc logs a warning. The warning is not needed, so this change
suppresses it by skipping that section when bash_completion.d is not
there.