bash: automatically source the checked-in bashrc
(I think my nix config used to do this. If so, I'm not sure when or why
the regression happened.)

This nix config pins my bash program, and it pins my bash configuration.
But before this commit, I was depending on implicit state to connect the
two together. Namely, I was depending on the existence of a ~/.bashrc
file on the host system to source ~/.nix-profile/config/bash/bashrc.sh.

I don't need that implicit state, and it is better to drop it. So I
have.