bump starship
v0.52.1 made starship faster in large git repos. This is a very welcome
change for me, since I sometimes interact with large repos on the
command line. github.com/NixOS/nixpkgs is a common one. I used to
hackily run `mv .git git` to force starship to stop treating it as a git
repo, lest I have to wait several seconds each time I complete a
command.

The new version is still noticeably slow, but it is at least tolerably
slow, which is a big improvement.