use vimPlugins.camelcasemotion from upstream
For the sake of shrinking the size of my personal environment, use the
upstream version of camelcasemotion.

This is a slightly different plugin. I was using this one before:

    https://github.com/vim-scripts/camelcasemotion

And the upstream one is this one:

    https://github.com/bkad/CamelCaseMotion

The latter is a fork of the former. The former has not been touched in
10 years, but the latter has had some development in that time. I'm not
really interested in that development, but I do like shrinking my
config.