init.vim: drop autosourcevimrc
This was a little hook to speed up iterating on my .vimrc: if I was
editing my .vimrc, then every time I saved the file it would source the
file again so I could see the changes right away.

This is no longer useful now that I do not edit my vimrc in place. And I
haven't had a need for it in years.