xonsh: drop prompt-toolkit
Two motivations for this:

- improve xonsh startup time

On my benchmark, xonsh starts up 300x slower than bash :(

The startup time on my machine is around 300ms, which is a noticeable
delay whenever I open a new shell. I quite dislike that, and want that
delay to go away.

On xonsh's own profiling (`xonsh --timings`), about 150ms of that
startup is spent initializing prompt-toolkit. Disabling prompt-toolkit
does seem to improve startup time, though weirdly by less than 150ms.

- disable a feature I didn't like

When you ctrl-c in xonsh, it blacks out your prompt and the text you had
previously entered. I find this kind of weird and don't really like it.
I searched around for ways to disable it and haven't found any yet. But
this feature is implemented in prompt-toolkit, and given that I want to
disable prompt-toolkit anyway for better startup time, this seems like a
fine way to disable the feature.