patch starship until starship/starship#2807 lands
starship does not have built-in support for xonsh yet, but there are
multiple open efforts to add it, including a PR I opened recently:

    https://github.com/starship/starship/pull/2807

Until that PR lands, I'm switching to using my patched version.

(I was already using a patched version on my mac, which I built locally.
Better to override it in the nix config so I automatically get it on
other machines too.)