simplify nvim config
A lot of this configuration is over a decade old. Much of it was made
obsolete by neovim (and perhaps even by vim) -- many of these items are
just the default now.  Some if it is for plugins I no longer have
installed.

And some of this was just normal code simplification: deleting
commented-out code and factoring out repetition.

I removed some `has('nvim')` and `has(<feature>)` guards, since I now
always use nvim, nvim does not conditionally compile features, and I pin
a very specific version of nvim through nix anyway.

I did some testing of the new config, but wasn't super careful to check
that it did everything I want. It's possible I'll want to revert a bit
of this in the future.