git: set default 'hub' protocol to ssh
See https://github.com/github/hub/issues/2937

My 'clone' command broke, for the reasons described in that issue. The
workaround is to set the default protocol to https or ssh. https (which
seems like it will become the new hub default) won't work for me since I
sometimes use this command to clone private repos. So I'm setting it to
ssh.