colorscheme: fix `colorscheme system`
The previous commit caused an interesting problem.

I switched from a rust expression to a python program specifically for
changing iterm's colors. This is because the previous rust expression
didn't work as well as I thought it did, and to properly fix it I needed
to call iterm2's API which only has bindings available for python. (I
could have created rust bindings, but that was too much work.)

This meant my script had a dependency on another script. Any time I add
a dependency on an external program, I want to capture it in the "deps"
attribute for the script, so that I know the script will have that
program available at runtime.

But in this case, I figured it didn't matter because the other script
(set-iterm2-colors) would also be installed in my profile anyway, so it
should work fine.

And indeed it did work fine when I invoked `colorscheme` directly on the
command line. But I also have my machine set up to invoke colorscheme
through launchctl, so that it can respond to colorscheme changes from
the OS.

The problem is that launchctl does not use the same PATH var that I have
set in my terminal. It uses a very restricted PATH instead:

	/usr/bin:/bin:/usr/sbin:/sbin

So when `colorscheme` was invoked from launchctl, my call to
`set-iterm2-colors` would fail and my colorscheme would break.

The solution I have settled on is to package `set-iterm2-colors` like
any other external program dependency.

I had to modify scripts.nix to make that happen, so that names in "deps"
could refer to other scripts. But I've done that and it works. Hooray!