drop `colorscheme restore-colors` concept
This subcommand used to have several responsibilities, but they were
slowly whittled down to just one thing: initialize the color theme in
kitty. But that can be done with an env var instead, which is less code
and (a teeny tiny bit) less startup time for every shell I launch.

So now that subcommand is gone and kitty just uses an env var on
startup.

For this to work I had to upgrade my `themed` function. Previously it
dropped all the outputs of a package except for its main binary. This
doesn't work for kitty since we need its terminfo output as well.

Also both `themed` and `fixGL` produced packages that did not have
`pname` set, but both depend on the `pname` of their input package,
which means they did not compose with each other. I fixed that too.