git: un-ignore some patterns
I lost some time today failing to find a search string because the file
I was looking for was in a directory called "tags". Sigh.

I think I added "tags" because of a vim tags feature? And I also think I
haven't used that feature in a long time, so this pattern is out.

Also dropping "/bazel-*" because I'm not using bazel right now, and
"*.swp" because my swap files should live in a separate directory now.