scripts/i.py: fail if update is a no-op
`nix flake update --commit-lock-file` unfortunately returns
_successfully_ even if it does not in fact update the lock file due to
dependencies being unlocked. So check for that happening and fail if it
does.

I want to both stream `nix flake update`'s output to stdout and also
programmatically read the output afterward, and the only succinct way I
found to do that was using the sh package, so this commit also switches
from subprocess to sh.