atuin: remove patch
Not needed now that I am using catppuccin. Previously I was using
solarized. Since solarized only has 16 colors, one of the colors has to
be the same as the background color. It happened that `atuin stats` used
that color.

catppuccin has more than 16 colors, so none of the 16 basic ansi colors
are the same as the background color.