-nix-index
I haven't been using this. Also it does not have a pname attribute,
which breaks a maintenance script I'm writing for this repo.

It doesn't have a pname attribute because the underlying rust
derivation, which _does_ have a pname attribute, is wrapped by a
symlinkJoin which (I think?) drops it:

https://github.com/NixOS/nixpkgs/blob/4f0dadbf38ee4cf4cc38cbc232b7708fddf965bc/pkgs/tools/package-management/nix-index/wrapper.nix