nvim: remove workaround
This reverts a previous commit:

    43e0fba nvim: bump to master to fix neovim/neovim#29550

neovim 0.11.4 is now released in nixpkgs. The comment in my workaround
says to wait until 0.12.0 is released, but according to this, 0.11.4 is
sufficient:

https://github.com/neovim/neovim/issues/29550#issuecomment-3181056461

I can no longer reproduce the original problem I was having, so I can't
confirm that this version works for me. But easy enough to switch back
to 0.12.0-dev if needed. (Which it probably won't be.)